Properly re-adjust DebugSequenceOverheadBenchmark
